src/test/java/de/techfak/gse/ysander/model/GridTest.java
package de.techfak.gse.ysander.model;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class GridTest {

    @Test
    void toFEN() {
        Grid grid = Grid.defaultGrid();

        assertEquals("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR", grid.toFEN());
    }

    @Test
    void fromFEN() {
        assertEquals(Grid.defaultGrid(), Grid.fromFEN("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR"));
        assertEquals("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR",
            Grid.fromFEN("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR").toFEN());
    }
}
